refactor: move backends out of grpc
