+++ pkg/backend/image/stablediffusion.go
