+++ pkg/backend/llm/gpt4all/gpt4all.go
